프로세스 모니터 구현중 Dispose패턴 구현하기

+++ ProjectKJServers/GameServer/MainUI/GameServer.cs
@@ -253,7 +253,7 @@ namespace GameServer
             DiskIOTextBox.Text = DiskIO.ToString();
         }
 
-        private void UpdateThreadUsage(int ThreadUsage)
+        private void UpdateThreadUsage(float ThreadUsage)
         {
             ThreadUsageTextBox.Text = ThreadUsage.ToString();
         }
